requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
apscheduler==3.10.4
python-multipart==0.0.6
pyjwt==2.12.1
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
